requirements.txt
# Core packages

numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0  # Optional alternative to PyTorch

# Geospatial
gdal>=3.2.0
rasterio>=1.2.0
geopandas>=0.10.0
shapely>=1.7.0
fiona>=1.8.0

# Image Processing
opencv-python>=4.5.0
scikit-image>=0.18.0
albumentations>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.60.0
pyyaml>=5.4.0
